Remove Mongoose in favor of NanoSQL
I want to use MongoDB but if I'm going to have a chance of finishing
this within the day I need to simplify the project a little. Hopefully
in the future I will be switching to MongoDB as although I like NanoSQL
the entire app is very scailable and it makes no sense to limit that
scailability by making the database be local to each instance of a api
service.